Use node_module caching solution only in production
