+++ webpack_config/production.js
@@ -6,6 +6,7 @@ const FaviconsWebpackPlugin = require('favicons-webpack-plugin');
 const SriPlugin = require('webpack-subresource-integrity');
 const common = require('./common');
 const config = require('./config');
+const { generateChunkName } = require('./utils');
 
 const IS_ELECTRON = !!process.env.BUILD_ELECTRON;
 
@@ -41,6 +42,19 @@ module.exports = merge.smart(common, {
         ]
       }
     ]
+  },
+
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        // Caching is only useful in production. This is is where we apply the node_module caching solution
+        vendorIndividual: {
+          enforce: true,
+          test: new RegExp(`[\\/]node_modules[\\/](${config.chunks.individual.join('|')})[\\/]`),
+          name: generateChunkName
+        }
+      }
+    }
   },
 
   plugins: [
